Move project.utils.highest_version to project.version_handling.highest_version
